refactor(web): modularize ListingDetail actions
- Extract useListingDetailActions hook
- Convert ListingDetail into orchestrator component (556 to 294 LOC)
- Preserve public behavior
